TradeGPT/backend/requirements.txt.txt
annotated-doc           
annotated-types         
anyio                    
bcrypt                  
certifi                 
cffi                     
charset-normalizer       
click                    
colorama                 
cryptography             
distro                  
dnspython                
email-validator          
fastapi                 
filetype                 
google-auth              
google-genai             
greenlet                 
h11                     
httpcore                 
httpx                    
idna                    
jsonpatch                
jsonpointer              
langchain                
langchain-classic        
langchain-core          
langchain-google-genai   
langchain-text-splitters 
langgraph                
langgraph-checkpoint     
langgraph-prebuilt       
langgraph-sdk            
langsmith                
orjson                   
ormsgpack              
packaging                
pip                     
psycopg2                 
psycopg2-binary         
pyasn1                   
pyasn1_modules          
pycparser               
pydantic                 
pydantic_core            
PyJWT                    
python-dotenv            
PyYAML                   
requests                 
requests-toolbelt        
sniffio                 
SQLAlchemy               
starlette                
tenacity               
typing_extensions       
typing-inspection       
urllib3                 
uuid_utils               
uvicorn                  
websockets
xxhash                   
zstandard                